+++ documentation/changelog.md
@@ -33,6 +33,9 @@
 
 
 ## Changelog:
+- 2019-10-08 Fixed handle functions with dots in name.
+- 2019-10-08 Added curved rectangles.
+- 2019-10-07 Added interactive shape - rectangle.
 - 2019-06-13 Fixed bug with responsive text entry using act.canvasdom.clientwidth.
 - 2019-05-22 Fixed bug in Objects when using a number as a object name.
 - 2019-05-09 Removed shapes begin, linewidth, stroke, fill, strokecolor, close, fillcolor.